qtwebkit: remove false dependency on leveldb
The qtwebkit configure test never succeeds in finding system leveldb
(ie qtwebkit internal copy of leveldb is always used) so leveldb is a
false dependency.

This change allows qtwebkit to be built in a disto which enables
security_flags.inc (it's otherwise blocked since the leveldb recipe
fails to build when security_flags.inc is used).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>