qtscript/qtwebkit: disable thumb for armv4 and armv5 only to fix segfault on armv7
We have a simple application containing a QWebView which segfaulted during
loading a page on a CortexA9 (armv7 thumb2). After enabling debug information
for qtwebkit and remote debug session the point of segfault was detected in

Source/JavaScriptCore/llint/LowLevelInterpreter.asm:1082
  callSlowPath(_llint_slow_path_new_func_exp)

Further digging around lead to similar source in:

* qtwebkit: Source/JavaScriptCore/assembler/ARMv7Assembler.h
* qtscript: src/3rdparty/javascriptcore/JavaScriptCore/assembler/ARMv7Assembler.h

In these files a class ARMThumbImmediate is declared and used in many places.
This lead me to the suspicion that compiligng for armv7 expects thumb support.

The following tests were performed:

* our application is working fine again
* qtscript and qtwebkit were build with
  'ARM_INSTRUCTION_SET = "thumb"' => TUNE_FEATURES = "arm armv5 thumb dsp"
  for qemuarm without issues

This is a partial/rebased revert of:
commit 932b32c1c9e9a3ac833c82ff9273683345056ec7
Author: Martin Jansa <Martin.Jansa@gmail.com>
Date:   Wed Apr 22 15:33:01 2015 +0200

    qtscript,qtwebking: disable thumb

    * qemuarm builds fail when thumb is enabled

    Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>

Signed-off-by: Andreas Müller <schnitzeltony@googlemail.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>